functions/predict_fabric_pattern/requirement.txt
tensorflow==2.18.0
numpy
Pillow